fishbowl/requirements.txt
BeautifulSoup==3.2.1
Django==1.7
Markdown==2.5.1
Pillow==2.6.1
PyYAML==3.11
Pygments==2.0.1
South==1.0.1
Unidecode==0.04.16
argparse==1.2.1
defusedxml==0.4.1
django-allauth==0.9.0
django-appconf==0.6
django-avatar==2.0
django-classy-tags==0.5.1
django-cms==3.0.6
django-filter==0.8
django-geoposition==0.2.2
django-mptt==0.6.1
django-oauth2-provider==0.2.6.1
django-registration==1.0
django-registration-redux==1.1
django-rest-authemail==0.1.4
django-sekizai==0.7
django-threadedcomments==0.9.0
djangocms-admin-style==0.2.2
djangorestframework==2.4.3
feedparser==5.1.3
gdata==2.0.18
goscalecms==0.5.6
html5lib==0.999
httplib2==0.9
lxml==3.4.0
oauth2==1.5.211
oauthlib==0.7.1
pbr==0.10.0
python-dateutil==2.2
python-openid==2.2.5
pytz==2014.7
requests==2.4.3
requests-oauthlib==0.4.2
shortuuid==0.4.2
simplejson==3.6.5
six==1.8.0
stevedore==1.1.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
xmltodict==0.9.0
